+++ Modules/SharePointDsc/DSCResources/MSFT_SPContentDatabase/MSFT_SPContentDatabase.psm1
@@ -4,14 +4,38 @@ function Get-TargetResource
     [OutputType([System.Collections.Hashtable])]
     param
     (
-        [parameter(Mandatory = $true)]  [System.String] $Name,
-        [parameter(Mandatory = $false)] [System.String] $DatabaseServer,
-        [parameter(Mandatory = $true)]  [System.String] $WebAppUrl,
-        [parameter(Mandatory = $false)] [System.Boolean] $Enabled,
-        [parameter(Mandatory = $false)] [System.UInt16] $WarningSiteCount,
-        [parameter(Mandatory = $false)] [System.UInt16] $MaximumSiteCount,
-        [parameter(Mandatory = $false)] [ValidateSet("Present","Absent")] [System.String] $Ensure = "Present",
-        [parameter(Mandatory = $false)] [System.Management.Automation.PSCredential] $InstallAccount
+        [parameter(Mandatory = $true)]
+        [System.String]
+        $Name,
+        
+        [parameter(Mandatory = $false)]
+        [System.String]
+        $DatabaseServer,
+        
+        [parameter(Mandatory = $true)]
+        [System.String]
+        $WebAppUrl,
+        
+        [parameter(Mandatory = $false)]
+        [System.Boolean]
+        $Enabled,
+        
+        [parameter(Mandatory = $false)]
+        [System.UInt16]
+        $WarningSiteCount,
+        
+        [parameter(Mandatory = $false)]
+        [System.UInt16]
+        $MaximumSiteCount,
+        
+        [parameter(Mandatory = $false)]
+        [ValidateSet("Present","Absent")]
+        [System.String]
+        $Ensure = "Present",
+        
+        [parameter(Mandatory = $false)]
+        [System.Management.Automation.PSCredential]
+        $InstallAccount
     )
 
     Write-Verbose -Message "Getting content database configuration settings"
@@ -74,25 +98,46 @@ function Set-TargetResource
     [CmdletBinding()]
     param
     (
-        [parameter(Mandatory = $true)]  [System.String] $Name,
-        [parameter(Mandatory = $false)] [System.String] $DatabaseServer,
-        [parameter(Mandatory = $true)]  [System.String] $WebAppUrl,
-        [parameter(Mandatory = $false)] [System.Boolean] $Enabled,
-        [parameter(Mandatory = $false)] [System.UInt16] $WarningSiteCount,
-        [parameter(Mandatory = $false)] [System.UInt16] $MaximumSiteCount,
-        [parameter(Mandatory = $false)] [ValidateSet("Present","Absent")] [System.String] $Ensure = "Present",
-        [parameter(Mandatory = $false)] [System.Management.Automation.PSCredential] $InstallAccount
+        [parameter(Mandatory = $true)]
+        [System.String]
+        $Name,
+        
+        [parameter(Mandatory = $false)]
+        [System.String]
+        $DatabaseServer,
+        
+        [parameter(Mandatory = $true)]
+        [System.String]
+        $WebAppUrl,
+        
+        [parameter(Mandatory = $false)]
+        [System.Boolean]
+        $Enabled,
+        
+        [parameter(Mandatory = $false)]
+        [System.UInt16]
+        $WarningSiteCount,
+        
+        [parameter(Mandatory = $false)]
+        [System.UInt16]
+        $MaximumSiteCount,
+        
+        [parameter(Mandatory = $false)]
+        [ValidateSet("Present","Absent")]
+        [System.String]
+        $Ensure = "Present",
+        
+        [parameter(Mandatory = $false)]
+        [System.Management.Automation.PSCredential]
+        $InstallAccount
     )
 
     Write-Verbose -Message "Setting content database configuration settings"
 
     Invoke-SPDSCCommand -Credential $InstallAccount `
-                        -Arguments @($PSBoundParameters,$PSScriptRoot) `
+                        -Arguments $PSBoundParameters `
                         -ScriptBlock {
         $params = $args[0]
-        $scriptRoot  = $args[1]
-
-        Import-Module (Join-Path $scriptRoot "..\..\Modules\SharePointDsc.ContentDatabase\SPContentDatabase.psm1" -Resolve)
 
         # Use Get-SPDatabase instead of Get-SPContentDatabase because the Get-SPContentDatabase
         # does not return disabled databases.
@@ -100,20 +145,24 @@ function Set-TargetResource
             $_.Type -eq "Content Database" -and $_.Name -eq $params.Name
         }
 
-        if ($params.Ensure -eq "Present") {
+        if ($params.Ensure -eq "Present")
+        {
             # Check if specified web application exists and throw exception when
             # this is not the case
             $webapp = Get-SPWebApplication | Where-Object -FilterScript {
                 $_.Url.Trim("/") -eq $params.WebAppUrl.Trim("/")
             }
 
-            if ($null -eq $webapp) {
+            if ($null -eq $webapp)
+            {
                 throw "Specified web application does not exist."
             }
 
             # Check if database exists
-            if ($null -ne $cdb) {
-                if ($cdb.Server -ne $params.DatabaseServer) {
+            if ($null -ne $cdb)
+            {
+                if ($cdb.Server -ne $params.DatabaseServer)
+                {
                     throw ("Specified database server does not match the actual database " + `
                            "server. This resource cannot move the database to a different " + `
                            "SQL instance.")
@@ -121,20 +170,65 @@ function Set-TargetResource
 
                 # Check and change attached web application.
                 # Dismount and mount to correct web application
-                if ($params.WebAppUrl.Trim("/") -ne $cdb.WebApplication.Url.Trim("/")) {
+                if ($params.WebAppUrl.Trim("/") -ne $cdb.WebApplication.Url.Trim("/"))
+                {
                     Dismount-SPContentDatabase $params.Name -Confirm:$false
 
-                    if ($params.ContainsKey("Enabled"))
+                    $newParams= @{}
+                    foreach ($param in $params.GetEnumerator())
                     {
-                        $enabled = $params.Enabled
+                        $skipParams = @("Enabled", "Ensure", "InstallAccount", "MaximumSiteCount", "WebAppUrl")
+
+                        if ($skipParams -notcontains $param.Key)
+                        {
+                            $newParams.$($param.Key) = $param.Value
+                        }
+                    
+                        if ($param.Key -eq "MaximumSiteCount")
+                        {
+                            $newParams.MaxSiteCount = $param.Value
+                        }
+
+                        if ($param.Key -eq "WebAppUrl")
+                        {
+                            $newParams.WebApplication = $param.Value
+                        }
+                    }
+
+                    try
+                    {
+                        $cdb = Mount-SPContentDatabase @newParams -ErrorAction Stop
+                    }
+                    catch
+                    {
+                        throw ("Error occurred while mounting content database. " + `
+                                "Content database is not mounted. " + `
+                                "Error details: $($_.Exception.Message)")
+                    }
+
+                    if ($cdb.Status -eq "Online")
+                    {
+                        $cdbenabled = $true
                     }
                     else
                     {
-                        $enabled = $true
+                        $cdbenabled = $false
                     }
                     
-                    $parameters = @{} + $params
-                    $cdb = Mount-SPDscContentDatabase $parameters $enabled
+                    if ($params.Enabled -ne $cdbenabled)
+                    {
+                        switch ($params.Enabled)
+                        {
+                            $true
+                            { 
+                                $cdb.Status = [Microsoft.SharePoint.Administration.SPObjectStatus]::Online
+                            }
+                            $false
+                            {
+                                $cdb.Status = [Microsoft.SharePoint.Administration.SPObjectStatus]::Disabled
+                            }
+                        }
+                    }
                 }
 
                 # Check and change database status
@@ -176,17 +270,61 @@ function Set-TargetResource
             else
             {
                 # Database does not exist, but should. Create/mount database
-                if ($params.ContainsKey("Enabled"))
+                $newParams= @{}
+                foreach ($param in $params.GetEnumerator())
                 {
-                    $enabled = $params.Enabled
+                    $skipParams = @("Enabled", "Ensure", "InstallAccount", "MaximumSiteCount", "WebAppUrl")
+
+                    if ($skipParams -notcontains $param.Key)
+                    {
+                        $newParams.$($param.Key) = $param.Value
+                    }
+                
+                    if ($param.Key -eq "MaximumSiteCount")
+                    {
+                        $newParams.MaxSiteCount = $param.Value
+                    }
+                    
+                    if ($param.Key -eq "WebAppUrl")
+                    {
+                        $newParams.WebApplication = $param.Value
+                    }
+                }
+
+                try
+                {
+                    $cdb = Mount-SPContentDatabase @newParams -ErrorAction Stop
+                }
+                catch
+                {
+                    throw ("Error occurred while mounting content database. " + `
+                            "Content database is not mounted. " + `
+                            "Error details: $($_.Exception.Message)")
+                }
+
+                if ($cdb.Status -eq "Online")
+                {
+                    $cdbenabled = $true
                 }
                 else
                 {
-                    $enabled = $true
+                    $cdbenabled = $false
                 }
                 
-                $parameters = @{} + $params
-                $cdb = Mount-SPDscContentDatabase $parameters $enabled
+                if ($params.Enabled -ne $cdbenabled)
+                {
+                    switch ($params.Enabled)
+                    {
+                        $true
+                        { 
+                            $cdb.Status = [Microsoft.SharePoint.Administration.SPObjectStatus]::Online
+                        }
+                        $false
+                        {
+                            $cdb.Status = [Microsoft.SharePoint.Administration.SPObjectStatus]::Disabled
+                        }
+                    }
+                }
             }
             $cdb.Update()
         }
@@ -208,17 +346,44 @@ function Test-TargetResource
     [OutputType([System.Boolean])]
     param
     (
-        [parameter(Mandatory = $true)]  [System.String] $Name,
-        [parameter(Mandatory = $false)] [System.String] $DatabaseServer,
-        [parameter(Mandatory = $true)]  [System.String] $WebAppUrl,
-        [parameter(Mandatory = $false)] [System.Boolean] $Enabled,
-        [parameter(Mandatory = $false)] [System.UInt16] $WarningSiteCount,
-        [parameter(Mandatory = $false)] [System.UInt16] $MaximumSiteCount,
-        [parameter(Mandatory = $false)] [ValidateSet("Present","Absent")] [System.String] $Ensure = "Present",
-        [parameter(Mandatory = $false)] [System.Management.Automation.PSCredential] $InstallAccount
+        [parameter(Mandatory = $true)]
+        [System.String]
+        $Name,
+        
+        [parameter(Mandatory = $false)]
+        [System.String]
+        $DatabaseServer,
+        
+        [parameter(Mandatory = $true)]
+        [System.String]
+        $WebAppUrl,
+        
+        [parameter(Mandatory = $false)]
+        [System.Boolean]
+        $Enabled,
+        
+        [parameter(Mandatory = $false)]
+        [System.UInt16]
+        $WarningSiteCount,
+        
+        [parameter(Mandatory = $false)]
+        [System.UInt16]
+        $MaximumSiteCount,
+        
+        [parameter(Mandatory = $false)]
+        [ValidateSet("Present","Absent")]
+        [System.String]
+        $Ensure = "Present",
+        
+        [parameter(Mandatory = $false)]
+        [System.Management.Automation.PSCredential]
+        $InstallAccount
     )
 
     Write-Verbose -Message "Testing content database configuration settings"
+
+    $PSBoundParameters.Ensure = $Ensure
+
     $CurrentValues = Get-TargetResource @PSBoundParameters
 
     if ($CurrentValues.DatabaseServer -ne $DatabaseServer)
@@ -229,7 +394,6 @@ function Test-TargetResource
         return $false
     }
 
-    $PSBoundParameters.Ensure = $Ensure
     return Test-SPDscParameterState -CurrentValues $CurrentValues `
                                     -DesiredValues $PSBoundParameters
 }