+++ Modules/SharePointDsc/DSCResources/MSFT_SPDiagnosticLoggingSettings/MSFT_SPDiagnosticLoggingSettings.psm1
@@ -306,11 +306,14 @@ function Test-TargetResource
     )
 
     Write-Verbose -Message "Testing diagnostic configuration settings"
+
     $CurrentValues = Get-TargetResource @PSBoundParameters
+
     if ($null -eq $CurrentValues) 
     { 
         return $false 
     }
+
     return Test-SPDscParameterState -CurrentValues $CurrentValues `
                                     -DesiredValues $PSBoundParameters
 }